requirements.txt
amqp==5.2.0
anyjson==0.3.3
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.88
botocore==1.34.88
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
Django==5.0.3
django-celery-beat==2.6.0
django-jet-reboot==1.3.7
django-nested-admin==4.0.2
django-rest-passwordreset==1.4.1
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flower==2.0.1
gunicorn==21.2.0
humanize==4.9.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.6
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.0
pillow==10.3.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-monkey-business==1.0.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
s3transfer==0.10.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tornado==6.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
